cambios en sintomas

+++ 04_TipoDeDato/DTSintoma.cs
@@ -6,18 +6,18 @@ namespace _4_TipoDeDato
 {
     public class DTSintoma
     {
-        private long id;
+        private int id;
         private string sintoma;
         private double coef;
         private double peligro;
 
-        public DTSintoma(long id, string sintoma)
+        public DTSintoma(int id, string sintoma)
         {
             this.id = id;
             this.sintoma = sintoma;
         }
 
-        public DTSintoma(long id, string sintoma, double coef)
+        public DTSintoma(int id, string sintoma, double coef)
         {
             this.id = id;
             this.sintoma = sintoma;
@@ -28,7 +28,7 @@ namespace _4_TipoDeDato
             DTSintoma sin = new DTSintoma(111,"hola");
 
         }
-        public long Id { get => id; set => id = value; }
+        public int Id { get => id; set => id = value; }
         public string Sintoma { get => sintoma; set => sintoma = value; }
         public double Coef { get => coef; set => coef = value; }
         public double Peligro { get => peligro; private set => peligro = value; }
